% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funksjoner.R
\name{beregn.arealvekt}
\alias{beregn.arealvekt}
\title{Calculate area weight}
\usage{
beregn.arealvekt(x, design, lokalitetsbredde = NA, rutebredde = 1)
}
\arguments{
\item{x}{Quadrate data}

\item{design}{Linje or Rose}

\item{lokalitetsbredde}{Width of locality}

\item{rutebredde}{with of quadrate (defaults to 1)}
}
\value{
area weighting of survey to calculate estimates
}
\description{
Calculate area weight
}
\keyword{internal}
