% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funksjoner.R
\name{beregn.dekning}
\alias{beregn.dekning}
\title{Calculate coverage}
\usage{
beregn.dekning(rutedata, artsdata, artsgruppe)
}
\arguments{
\item{rutedata}{quadrate data}

\item{artsdata}{species data}

\item{artsgruppe}{Species}
}
\value{
coverage
}
\description{
Calculate coverage
}
