% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funksjoner.R
\name{beregn.forekomstareal}
\alias{beregn.forekomstareal}
\title{Calculate the Occurrence area}
\usage{
beregn.forekomstareal(
  transektdata,
  utvalgt.lokalitet,
  year,
  design,
  lokalitetsbredde,
  rutebredde = 1,
  forekomst_transekt = "Dragehode",
  forekomst_avstand = "Forekomst dragehode (m)"
)
}
\arguments{
\item{transektdata}{transect data}

\item{utvalgt.lokalitet}{chosen locality}

\item{year}{year}

\item{design}{Linje or Rose}

\item{lokalitetsbredde}{Width of locality}

\item{rutebredde}{with of quadrate (defaults to 1)}

\item{forekomst_transekt}{"Dragehode" column in data}

\item{forekomst_avstand}{"Forekomst dragehode (m)" column in data}
}
\description{
Calculate the Occurrence area
}
\keyword{internal}
