% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funksjoner.R
\name{beregn.vekstrate}
\alias{beregn.vekstrate}
\title{Calculate growth rate}
\usage{
beregn.vekstrate(rutedata, popvar, idvar)
}
\arguments{
\item{rutedata}{quadrate data}

\item{popvar}{population variable (Total Dragehode )}

\item{idvar}{Id of the quadrate (RuteID)}
}
\value{
growthrate estimates
}
\description{
Calculate growth rate
}
