% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funksjoner.R
\name{boot.pop}
\alias{boot.pop}
\title{Bootstrap population}
\usage{
boot.pop(tetthet, forekomstareal, nboot = 2000)
}
\arguments{
\item{tetthet}{density}

\item{forekomstareal}{Occurrence area}

\item{nboot}{number of bootstrap iterations (defaults to 2000)}
}
\value{
bootstraped estimates of population per age stage
}
\description{
Bootstrap population
}
\keyword{internal}
