% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funksjoner.R
\name{ggplot.popstruktur}
\alias{ggplot.popstruktur}
\title{GGplot version of plot.popstruktur}
\usage{
ggplot.popstruktur(popstr)
}
\arguments{
\item{popstr}{population structure}
}
\value{
plots of population structure
}
\description{
GGplot version of plot.popstruktur
}
