% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funksjoner.R
\name{plot.gruppetrender}
\alias{plot.gruppetrender}
\title{Base plot group trends}
\usage{
\method{plot}{gruppetrender}(
  lokalitetsdata,
  gruppevariabel,
  lokalitetsestimater,
  reverser = FALSE,
  lokalitetsnavn,
  lokfarge,
  regsymbol,
  natlinje
)
}
\arguments{
\item{lokalitetsdata}{locality data}

\item{gruppevariabel}{grouping variable}

\item{lokalitetsestimater}{local estimates data}

\item{reverser}{Reverse the axis? default is FALSE}

\item{lokalitetsnavn}{locality name}

\item{lokfarge}{locality colour}

\item{regsymbol}{Regional symbol}

\item{natlinje}{linetype for naturtype}
}
\value{
baseplots of group trends
}
\description{
Base plot group trends
}
