% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funksjoner.R
\name{plot.popstruktur}
\alias{plot.popstruktur}
\title{Base plot population structure}
\usage{
\method{plot}{popstruktur}(popstr)
}
\arguments{
\item{popstr}{population structure data}
}
\description{
Base plot population structure
}
